refactor: change location of cardStore
