+++ src/components/cardsList.tsx
@@ -2,7 +2,7 @@ import React, { type ReactNode } from 'react';
 import type { CharacterType } from '../models/types';
 import { Card } from './card';
 import { useNavigate } from 'react-router-dom';
-import { useCardStore } from '../services/cardStore';
+import { useSelectedIds } from '../store/cardStore';
 import { SelectedItemsPanel } from './selectedItemsPanel';
 
 interface ListProps {
@@ -12,7 +12,7 @@ interface ListProps {
 
 export const CardsList: React.FC<ListProps> = ({ results, onClick }): ReactNode => {
   const navigate = useNavigate();
-  const selectedIds = useCardStore((s) => s.selectedIds);
+  const selectedIds = useSelectedIds();
   return (
     <div className="flex flex-wrap justify-center gap-4" onClick={onClick}>
       {results.map((char) => (
